feat: add propertiesAggregatorService (#289)
Description
Add a new PropertiesAggregatorService which can aggregate multiple property queries on a contract into a single call, with the limitation that there can be no arguments in the properties. This can be used, for example, for fetching multiple properties of a vault at once, while being flexible for adding/removing the properties queried in the future.

Corresponding PR of where the smart contract functionality has been added to lens - yearn/yearn-lens#55

Motivation and Context
Adding more functionality to the SDK for rendering data about vaults/strategies. This will be followed by adding a function to the vaults interface that supports fetching information about a vault in one call, e.g. name, fees, totalAssets etc

How Has This Been Tested?
Added unit tests